+++ R/quasiquotation.R
@@ -70,12 +70,23 @@
 #' @section Life cycle:
 #'
 #' * Calling `UQ()` and `UQS()` with the rlang namespace qualifier is
-#'   soft-deprecated as of rlang 0.2.0. Just use the unqualified forms
-#'   instead.
+#'   deprecated as of rlang 0.3.0. Just use the unqualified forms
+#'   instead:
+#'
+#'   ```
+#'   # Bad
+#'   rlang::expr(mean(rlang::UQ(var) * 100))
+#'
+#'   # Ok
+#'   rlang::expr(mean(UQ(var) * 100))
+#'
+#'   # Good
+#'   rlang::expr(mean(!!var * 100))
+#'   ```
 #'
 #'   Supporting namespace qualifiers complicates the implementation of
 #'   unquotation and is misleading as to the nature of unquoting
-#'   operators (these are syntactic operators that operates at
+#'   operators (which are syntactic operators that operates at
 #'   quotation-time rather than function calls at evaluation-time).
 #'
 #' * `UQ()` and `UQS()` were soft-deprecated in rlang 0.2.0 in order