+++ src/internal/expr-interp.c
@@ -72,15 +72,33 @@ void signal_uqs_soft_deprecation() {
 }
 
 void signal_namespaced_uq_deprecation() {
-  signal_soft_deprecation(
-    "Prefixing `UQ()` with a namespace is soft-deprecated as of rlang 0.2.0. "
-    "Please use the unprefixed form instead."
+  r_warn(
+    "Prefixing `UQ()` with the rlang namespace is deprecated as of rlang 0.3.0.\n"
+    "Please use the non-prefixed form or `!!` instead.\n"
+    "\n"
+    "  # Bad:\n"
+    "  rlang::expr(mean(rlang::UQ(var) * 100))\n"
+    "\n"
+    "  # Ok:\n"
+    "  rlang::expr(mean(UQ(var) * 100))\n"
+    "\n"
+    "  # Good:\n"
+    "  rlang::expr(mean(!!var * 100))\n"
   );
 }
 void signal_namespaced_uqs_deprecation() {
-  signal_soft_deprecation(
-    "Prefixing `UQS()` with a namespace is soft-deprecated as of rlang 0.2.0. "
-    "Please use the unprefixed form instead."
+  r_warn(
+    "Prefixing `UQS()` with the rlang namespace is deprecated as of rlang 0.3.0.\n"
+    "Please use the non-prefixed form or `!!!` instead.\n"
+    "\n"
+    "  # Bad:\n"
+    "  rlang::expr(mean(rlang::UQS(args)))\n"
+    "\n"
+    "  # Ok:\n"
+    "  rlang::expr(mean(UQS(args)))\n"
+    "\n"
+    "  # Good:\n"
+    "  rlang::expr(mean(!!!args))\n"
   );
 }